Updated error code for exit hello and norminette a few other codes.
